refactor: storage service package location
