+++ src/main/java/com/farneser/cloudfilestorage/controller/SearchController.java
@@ -2,7 +2,7 @@ package com.farneser.cloudfilestorage.controller;
 
 import com.farneser.cloudfilestorage.exception.EmptyQueryException;
 import com.farneser.cloudfilestorage.exception.InternalServerException;
-import com.farneser.cloudfilestorage.service.StorageService;
+import com.farneser.cloudfilestorage.service.storage.StorageService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
